+++ package.json
@@ -16,6 +16,7 @@
     "react-dom": "^16.8.3",
     "react-highlight": "^0.12.0",
     "react-scripts": "2.1.5",
+    "react-router-bootstrap": "^0.25.0",
     "react-router-dom": "^5.0.0",
     "styled-components": "^4.2.0"
   },
