+++ src/App/Common/Header.js
@@ -1,22 +1,32 @@
-import React, { Component } from 'react';
-import { Link } from 'react-router-dom'
+import React, { Component } from "react";
+// import Nav from "react-bootstrap/Nav";
+// import Navbar from "react-bootstrap/Navbar";
+
+import { Nav, Navbar, NavItem } from "react-bootstrap";
+import { LinkContainer } from "react-router-bootstrap";
 
 class Header extends Component {
-
-    render() {
-      return (
-            <header>
-                <nav>
-                    <ul>
-                    <li><Link to='/'>Home</Link></li>
-                    <li><Link to='/tutorial'>Tutorial</Link></li>
-                    <li><Link to='/events'>Events</Link></li>
-                    <li><Link to='/login'>Login</Link></li>
-                    </ul>
-                </nav>
-            </header>
-      );
-    }
+  render() {
+    return (
+      <Navbar collapseOnSelect expand="lg" bg="dark" variant="dark">
+        <Navbar.Brand href="#home">Brand Bar</Navbar.Brand>
+        <Navbar.Toggle aria-controls="responsive-navbar-nav" />
+        <Navbar.Collapse id="responsive-navbar-nav">
+          <Nav className="mr-auto">
+          <LinkContainer to="/home">
+              <NavItem>Home</NavItem>
+            </LinkContainer>
+            <LinkContainer to="/tutorial">
+              <NavItem>Tutorial</NavItem>
+            </LinkContainer>
+            <LinkContainer to="/events">
+              <NavItem>Events</NavItem>
+            </LinkContainer>
+          </Nav>
+        </Navbar.Collapse>
+      </Navbar>
+    );
   }
+}
 
 export default Header;
