--- src/App/Pages/Events/EventsMain.js
@@ -1,14 +1,0 @@
-import React, { Component } from 'react';
-
-class EventsMain extends Component {
-
-  render() {
-    return (
-      <div>
-        Events Page
-      </div>
-    );
-  }
-}
-
-export default EventsMain;
